Move class `FloatRoundedRect::Radii` into new class `CornerRadii` https://bugs.webkit.org/show_bug.cgi?id=305447 rdar://168117393
Reviewed by Abrar Rahman Protyasha and Wenson Hsieh.

Move `FloatRoundedRect::Radii` into new class `CornerRadii` so that it can
be used without importing `FloatRoundedRect`, and so that serialization
will be simpler (which will be useful in a later patch).

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/inspector/InspectorOverlay.cpp:
(WebCore::InspectorOverlay::drawRulers):
* Source/WebCore/inspector/InspectorOverlayLabel.cpp:
(WebCore::InspectorOverlayLabel::draw):
* Source/WebCore/page/DebugPageOverlays.cpp:
(WebCore::drawCheckbox):
(WebCore::InteractionRegionOverlay::drawSettings):
* Source/WebCore/platform/DictationCaretAnimator.cpp:
(WebCore::DictationCaretAnimator::expandedCaretRect const):
* Source/WebCore/platform/OpacityCaretAnimator.cpp:
(WebCore::OpacityCaretAnimator::paint const):
* Source/WebCore/platform/cocoa/DragImageCocoa.mm:
(WebCore::createDragImageForLink):
* Source/WebCore/platform/graphics/CornerRadii.cpp: Added.
(WebCore::CornerRadii::isZero const):
(WebCore::CornerRadii::hasEvenCorners const):
(WebCore::CornerRadii::isUniformCornerRadius const):
(WebCore::CornerRadii::scale):
(WebCore::CornerRadii::expand):
(WebCore::CornerRadii::expandEvenIfZero):
(WebCore::operator<<):
* Source/WebCore/platform/graphics/CornerRadii.h: Added.
(WebCore::CornerRadii::CornerRadii):
(WebCore::CornerRadii::setTopLeft):
(WebCore::CornerRadii::setTopRight):
(WebCore::CornerRadii::setBottomLeft):
(WebCore::CornerRadii::setBottomRight):
(WebCore::CornerRadii::topLeft const):
(WebCore::CornerRadii::topRight const):
(WebCore::CornerRadii::bottomLeft const):
(WebCore::CornerRadii::bottomRight const):
(WebCore::CornerRadii::expand):
(WebCore::CornerRadii::shrink):
* Source/WebCore/platform/graphics/FloatRoundedRect.cpp:
(WebCore::FloatRoundedRect::FloatRoundedRect):
(WebCore::operator<<):
(WebCore::FloatRoundedRect::Radii::isZero const): Deleted.
(WebCore::FloatRoundedRect::Radii::hasEvenCorners const): Deleted.
(WebCore::FloatRoundedRect::Radii::isUniformCornerRadius const): Deleted.
(WebCore::FloatRoundedRect::Radii::scale): Deleted.
(WebCore::FloatRoundedRect::Radii::expand): Deleted.
(WebCore::FloatRoundedRect::Radii::expandEvenIfZero): Deleted.
* Source/WebCore/platform/graphics/FloatRoundedRect.h:
(WebCore::FloatRoundedRect::radii const):
(WebCore::FloatRoundedRect::setRadii):
(WebCore::calcBorderRadiiConstraintScaleFor):
(WebCore::FloatRoundedRect::Radii::Radii): Deleted.
(WebCore::FloatRoundedRect::Radii::setTopLeft): Deleted.
(WebCore::FloatRoundedRect::Radii::setTopRight): Deleted.
(WebCore::FloatRoundedRect::Radii::setBottomLeft): Deleted.
(WebCore::FloatRoundedRect::Radii::setBottomRight): Deleted.
(WebCore::FloatRoundedRect::Radii::topLeft const): Deleted.
(WebCore::FloatRoundedRect::Radii::topRight const): Deleted.
(WebCore::FloatRoundedRect::Radii::bottomLeft const): Deleted.
(WebCore::FloatRoundedRect::Radii::bottomRight const): Deleted.
(WebCore::FloatRoundedRect::Radii::expand): Deleted.
(WebCore::FloatRoundedRect::Radii::shrink): Deleted.
* Source/WebCore/platform/graphics/LayoutRoundedRect.cpp:
(WebCore::LayoutRoundedRect::pixelSnappedRoundedRectForPainting const):
* Source/WebCore/platform/graphics/NativeImageSource.cpp:
* Source/WebCore/platform/graphics/PathUtilities.cpp:
(WebCore::PathUtilities::pathWithShrinkWrappedRects):
* Source/WebCore/platform/graphics/PathUtilities.h:
* Source/WebCore/platform/graphics/ShadowBlur.cpp:
(WebCore::computeSliceSizesFromRadii):
(WebCore::ShadowBlur::templateSize const):
(WebCore::ShadowBlur::drawLayerPieces):
(WebCore::ShadowBlur::drawLayerPiecesAndFillCenter):
* Source/WebCore/platform/graphics/ShadowBlur.h:
* Source/WebCore/platform/graphics/adwaita/SliderTrackAdwaita.cpp:
(WebCore::SliderTrackAdwaita::draw):
* Source/WebCore/platform/graphics/avfoundation/FormatDescriptionUtilities.cpp:
* Source/WebCore/platform/graphics/cairo/PathCairo.cpp:
(WebCore::PathCairo::add):
* Source/WebCore/platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContextCG::fillRoundedRectImpl):
* Source/WebCore/platform/graphics/cg/PathCG.cpp:
(WebCore::addToCGPath):
* Source/WebCore/platform/graphics/mac/controls/ControlMac.mm:
(WebCore::ControlMac::drawListButton):
* Source/WebCore/platform/graphics/skia/PathSkia.cpp:
(WebCore::PathSkia::add):
* Source/WebCore/platform/ios/DragImageIOS.mm:
(WebCore::createDragImageForColor):
* Source/WebCore/rendering/OutlinePainter.cpp:
(WebCore::adjustedRadiiForHuggingCurve):
(WebCore::OutlinePainter::pathWithShrinkWrappedRects):
* Source/WebCore/rendering/RenderTheme.cpp:
(WebCore::RenderTheme::paintSystemPreviewBadge):
* Source/WebCore/rendering/TextBoxPainter.cpp:
(WebCore::TextBoxPainter::paintBackgroundFillForRange):
(WebCore::radiiForUnderline):
(WebCore::trimRadii):
(WebCore::TextBoxPainter::fillCompositionUnderline const):
(WebCore::TextBoxPainter::paintCompositionUnderline):
* Source/WebCore/rendering/TextBoxPainter.h:
* Source/WebCore/rendering/cocoa/RenderThemeCocoa.mm:
(WebCore::roundedShape):
(WebCore::RenderThemeCocoa::paintColorWellForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintColorWellSwatchForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintInnerSpinButtonForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintMeterForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintProgressBarForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintSliderTrackForVectorBasedControls):
(WebCore::RenderThemeCocoa::paintPlatformResizerForVectorBasedControls):
* Source/WebCore/rendering/ios/RenderThemeIOS.mm:
(WebCore::RenderThemeIOS::paintSliderTrack):
(WebCore::RenderThemeIOS::paintProgressBar):
(WebCore::RenderThemeIOS::paintCheckbox):
(WebCore::RenderThemeIOS::paintRadio):
(WebCore::RenderThemeIOS::paintMeter):
(WebCore::RenderThemeIOS::paintSliderTicks):
* Source/WebCore/rendering/mac/RenderThemeMac.mm:
(WebCore::paintAttachmentIconBackground):
(WebCore::paintAttachmentProgress):
* Source/WebCore/rendering/shapes/LayoutShape.cpp:
(WebCore::LayoutShape::createShape):
* Source/WebCore/style/values/borders/StyleBorderRadius.cpp:
(WebCore::Style::CornerRadii>::operator):
(WebCore::Style::FloatRoundedRect::Radii>::operator): Deleted.
* Source/WebCore/style/values/borders/StyleBorderRadius.h:
* Source/WebCore/style/values/shapes/StyleInsetFunction.cpp:
(WebCore::Style::PathComputation<Inset>::operator):
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _updateTapHighlight]):
* Source/WebKit/UIProcess/ios/WKTapHighlightView.h:
* Source/WebKit/UIProcess/ios/WKTapHighlightView.mm:
(-[WKTapHighlightView setCornerRadii:]):

Canonical link: https://commits.webkit.org/305588@main
