+++ Source/WebCore/platform/DictationCaretAnimator.cpp
@@ -284,7 +284,7 @@ void DictationCaretAnimator::fillCaretTail(const FloatRect& rect, GraphicsContex
 FloatRoundedRect DictationCaretAnimator::expandedCaretRect(const FloatRect& rect, bool fillTail) const
 {
     if (m_initialScale <= 0.f && fillTail)
-        return FloatRoundedRect { rect, FloatRoundedRect::Radii { 1.f } };
+        return FloatRoundedRect { rect, CornerRadii { 1.f } };
 
     auto extraScaleFactor = 1.f;
     auto pulseExpansion = 1.f;
@@ -301,7 +301,7 @@ FloatRoundedRect DictationCaretAnimator::expandedCaretRect(const FloatRect& rect
     float verticalPulseExpansion = pulseExpansion * (1.f + 3.f * (extraScaleFactor - 1.f)) - 1.f;
     FloatRect expandedRect = rect;
     expandedRect.expand(FloatBoxExtent { verticalPulseExpansion, horizontalPulseExpansion, verticalPulseExpansion, horizontalPulseExpansion });
-    return FloatRoundedRect { expandedRect, FloatRoundedRect::Radii(1.f + std::max(0.f, .5f * horizontalPulseExpansion), 1.f + std::max(0.f, .5f * horizontalPulseExpansion)) };
+    return FloatRoundedRect { expandedRect, CornerRadii(1.f + std::max(0.f, .5f * horizontalPulseExpansion), 1.f + std::max(0.f, .5f * horizontalPulseExpansion)) };
 }
 
 void DictationCaretAnimator::paint(GraphicsContext& context, const FloatRect& rect, const Color& caretColor, const LayoutPoint& paintOffset) const
