+++ Source/WebCore/platform/OpacityCaretAnimator.cpp
@@ -121,7 +121,7 @@ void OpacityCaretAnimator::paint(GraphicsContext& context, const FloatRect& rect
     if (caretColor != Color::transparentBlack)
         caretColorWithOpacity = caretColor.colorWithAlpha(caretPresentationProperties.opacity);
 
-    context.fillRoundedRect(FloatRoundedRect { rect, FloatRoundedRect::Radii { 1.0 } }, caretColorWithOpacity);
+    context.fillRoundedRect(FloatRoundedRect { rect, CornerRadii { 1.0 } }, caretColorWithOpacity);
 }
 
 LayoutRect OpacityCaretAnimator::caretRepaintRectForLocalRect(LayoutRect repaintRect) const
