+++ Source/WebCore/platform/graphics/FloatRoundedRect.h
@@ -30,6 +30,7 @@
 #ifndef FloatRoundedRect_h
 #define FloatRoundedRect_h
 
+#include <WebCore/CornerRadii.h>
 #include <WebCore/FloatRect.h>
 #include <WebCore/FloatSize.h>
 #include <WebCore/LayoutRoundedRect.h>
@@ -47,85 +48,19 @@ class Path;
 class FloatRoundedRect {
     WTF_MAKE_TZONE_ALLOCATED_EXPORT(FloatRoundedRect, WEBCORE_EXPORT);
 public:
-    class Radii {
-        WTF_MAKE_TZONE_ALLOCATED(Radii);
-    public:
-        Radii() = default;
-        Radii(const FloatSize& topLeft, const FloatSize& topRight, const FloatSize& bottomLeft, const FloatSize& bottomRight)
-            : m_topLeft(topLeft)
-            , m_topRight(topRight)
-            , m_bottomLeft(bottomLeft)
-            , m_bottomRight(bottomRight)
-        {
-        }
-
-        Radii(const LayoutRoundedRect::Radii& intRadii)
-            : m_topLeft(intRadii.topLeft())
-            , m_topRight(intRadii.topRight())
-            , m_bottomLeft(intRadii.bottomLeft())
-            , m_bottomRight(intRadii.bottomRight())
-        {
-        }
-
-        explicit Radii(float uniformRadius)
-            : m_topLeft(uniformRadius, uniformRadius)
-            , m_topRight(uniformRadius, uniformRadius)
-            , m_bottomLeft(uniformRadius, uniformRadius)
-            , m_bottomRight(uniformRadius, uniformRadius)
-        {
-        }
-
-        explicit Radii(float uniformRadiusWidth, float uniformRadiusHeight)
-            : m_topLeft(uniformRadiusWidth, uniformRadiusHeight)
-            , m_topRight(uniformRadiusWidth, uniformRadiusHeight)
-            , m_bottomLeft(uniformRadiusWidth, uniformRadiusHeight)
-            , m_bottomRight(uniformRadiusWidth, uniformRadiusHeight)
-        {
-        }
-
-        void setTopLeft(const FloatSize& size) { m_topLeft = size; }
-        void setTopRight(const FloatSize& size) { m_topRight = size; }
-        void setBottomLeft(const FloatSize& size) { m_bottomLeft = size; }
-        void setBottomRight(const FloatSize& size) { m_bottomRight = size; }
-        const FloatSize& topLeft() const { return m_topLeft; }
-        const FloatSize& topRight() const { return m_topRight; }
-        const FloatSize& bottomLeft() const { return m_bottomLeft; }
-        const FloatSize& bottomRight() const { return m_bottomRight; }
-
-        bool isZero() const;
-        bool hasEvenCorners() const;
-        bool isUniformCornerRadius() const; // Including no radius.
-
-        void scale(float factor);
-        void scale(float horizontalFactor, float verticalFactor);
-        void expandEvenIfZero(float size);
-        void expand(float topWidth, float bottomWidth, float leftWidth, float rightWidth);
-        void expand(float size) { expand(size, size, size, size); }
-        void shrink(float topWidth, float bottomWidth, float leftWidth, float rightWidth) { expand(-topWidth, -bottomWidth, -leftWidth, -rightWidth); }
-        void shrink(float size) { shrink(size, size, size, size); }
-
-        friend bool operator==(const Radii&, const Radii&) = default;
-
-    private:
-        FloatSize m_topLeft;
-        FloatSize m_topRight;
-        FloatSize m_bottomLeft;
-        FloatSize m_bottomRight;
-    };
-
-    WEBCORE_EXPORT explicit FloatRoundedRect(const FloatRect& = FloatRect(), const Radii& = Radii());
+    WEBCORE_EXPORT explicit FloatRoundedRect(const FloatRect& = FloatRect(), const CornerRadii& = CornerRadii());
     WEBCORE_EXPORT FloatRoundedRect(const FloatRect&, const FloatSize& topLeft, const FloatSize& topRight, const FloatSize& bottomLeft, const FloatSize& bottomRight);
     explicit FloatRoundedRect(const LayoutRoundedRect&);
     FloatRoundedRect(float x, float y, float width, float height);
 
     const FloatRect& rect() const { return m_rect; }
-    const Radii& radii() const { return m_radii; }
+    const CornerRadii& radii() const { return m_radii; }
     bool isRounded() const { return !m_radii.isZero(); }
     bool isEmpty() const { return m_rect.isEmpty(); }
 
     void setRect(const FloatRect& rect) { m_rect = rect; }
     void setLocation(FloatPoint location) { m_rect.setLocation(location); }
-    void setRadii(const Radii& radii) { m_radii = radii; }
+    void setRadii(const CornerRadii& radii) { m_radii = radii; }
 
     void move(const FloatSize& size) { m_rect.move(size); }
     void inflate(float size) { m_rect.inflate(size);  }
@@ -167,10 +102,10 @@ public:
 
 private:
     FloatRect m_rect;
-    Radii m_radii;
+    CornerRadii m_radii;
 };
 
-inline float calcBorderRadiiConstraintScaleFor(const FloatRect& rect, const FloatRoundedRect::Radii& radii)
+inline float calcBorderRadiiConstraintScaleFor(const FloatRect& rect, const CornerRadii& radii)
 {
     // Constrain corner radii using CSS3 rules:
     // http://www.w3.org/TR/css3-background/#the-border-radius
