+++ Source/WebCore/platform/graphics/NativeImageSource.cpp
@@ -25,6 +25,7 @@
 
 #include "config.h"
 #include "NativeImageSource.h"
+#include <wtf/text/TextStream.h>
 
 namespace WebCore {
