+++ Source/WebCore/platform/graphics/cairo/PathCairo.cpp
@@ -324,7 +324,7 @@ void PathCairo::add(PathRoundedRect roundedRect)
 void PathCairo::add(PathContinuousRoundedRect continuousRoundedRect)
 {
     // Continuous rounded rects are unavailable. Paint a normal rounded rect instead.
-    add(PathRoundedRect { FloatRoundedRect { continuousRoundedRect.rect, FloatRoundedRect::Radii { continuousRoundedRect.cornerWidth, continuousRoundedRect.cornerHeight } }, PathRoundedRect::Strategy::PreferNative });
+    add(PathRoundedRect { FloatRoundedRect { continuousRoundedRect.rect, CornerRadii { continuousRoundedRect.cornerWidth, continuousRoundedRect.cornerHeight } }, PathRoundedRect::Strategy::PreferNative });
 }
 
 void PathCairo::add(PathCloseSubpath)
