+++ Source/WebCore/style/values/borders/StyleBorderRadius.h
@@ -79,8 +79,8 @@ template<> struct CSSValueConversion<BorderRadiusValue> { auto operator()(Builde
 
 // MARK: - Evaluation
 
-template<> struct Evaluation<BorderRadius, FloatRoundedRect::Radii> {
-    auto operator()(const BorderRadius&, FloatSize, ZoomNeeded) -> FloatRoundedRect::Radii;
+template<> struct Evaluation<BorderRadius, CornerRadii> {
+    auto operator()(const BorderRadius&, FloatSize, ZoomNeeded) -> CornerRadii;
 };
 template<> struct Evaluation<BorderRadius, LayoutRoundedRect::Radii> {
     auto operator()(const BorderRadius&, LayoutSize, ZoomNeeded) -> LayoutRoundedRect::Radii;
